Update UARPIII defconfig
I've given up trying to add EPCQL flash loading to the mess that is the SPL codebase. Instead we will just load an image from the MMC which will be a completely custom second stage bootloader. This is given the upper 64kB of on-chip RAM (SPL gets lower 192kB).
The custom second stage will then do everything needed to load the image from the QSPI into the lower 192kB of on-chip RAM (SPL has finished with it by that point). Much simpler.