Update UARPIII Configuration
Reduce the padding of the SPL image to make it smaller - no point programming 120kB of 0's!

Also move the stack/heap address to make a bit more heap room just in case it is needed for parsing the device tree.

Finally correct the SPL_REMOVE_PROPS setting to remove "clock-names" rather than "clocks" - we need to keep the clock handles, it's the names that we don't care about.