Set Maximum Boot Memory Length
This is limited as we only have on-chip RAM to play with. The SPL is given up to 192kB (not that much needed, I know), which leaves 64kB for our recovery image. The odd split allows the recovery image to then load the main application over the top of the SPL.