Add correct SPL_SIZE_LIMIT and make LSC preloader smaller
Set the correct size limit for the Cyclone V - remember the boot ROM reserves 4kB! Somebody messed up and set a default that allows the preloader to fill this space, which causes the boot ROM to crash. Change also some LSC options to make the preloader small enough to fit.